requirements.txt
m3u8==2.0.0
requests==2.28.1
vk_api==11.9.9
pycryptodome==3.15.0
beautifulsoup4==4.11.1